# add --unstable-feature=regex_link_parsing for a 12% resolve speedup