### TODO
- [ ] complete the regex parsing implementation and ensure it passes all the tests

### Problem

In performing some profiling (with [`py-spy`](https://pypi.org/project/py-spy/)) for #8448 and #8480, I found that calling `pip._internal.models.wheel.Wheel.supported()` and `pip._internal.index.collector.parse_links()` (specifically `html5lib.parse()`) took up almost the entire time of finding candidates.

I previously worked on this method in #7729, which was an optimization that [`pex`](github.com/pantsbuild/pex/) version 1 had made before moving to invoking pip as a subprocess in pex versions >=2.

### Solution

- Create a new option `--unstable-feature=regex_link_parsing` to use an experimental parser that uses regular expressions instead of `html5lib`.

### Result
A 12% decrease in the mean time to download `tensorflow==1.14.0` and all of its dependencies over 6 runs.
```bash
> for i in $(seq 6); do rm -rf tmp/ ~/Library/Caches/pip && time pip download --unstable-feature=resolver --unstable-feature=regex_link_parsing -d tmp/ tensorflow==1.14.0 &>/dev/null; done
13.11s user 3.61s system 76% cpu 21.752 total
12.97s user 3.55s system 85% cpu 19.358 total
12.92s user 3.41s system 85% cpu 19.136 total
12.84s user 3.38s system 85% cpu 19.076 total
12.72s user 3.36s system 86% cpu 18.623 total
12.90s user 3.39s system 85% cpu 18.989 total
> for i in $(seq 6); do rm -rf tmp/ ~/Library/Caches/pip && time pip download --unstable-feature=resolver -d tmp/ tensorflow==1.14.0 &>/dev/null; done
15.32s user 4.00s system 82% cpu 23.333 total
14.87s user 3.48s system 88% cpu 20.844 total
14.83s user 3.60s system 86% cpu 21.310 total
15.21s user 3.62s system 88% cpu 21.372 total
15.41s user 3.78s system 82% cpu 23.381 total
14.83s user 3.40s system 86% cpu 21.020 total
```