@pradyunsg 2020-06-23
I *really* don't think we should be adding regex-based link parsing to pip -- HTML is based on a context free grammar, and regex can only model/match regular languages (which is strictly a subset of CFGs).

There's a lot of "don't use regex for processing [X]HTML" content on the internet (for good reason). Here's one that comes to mind easily: https://blog.codinghorror.com/parsing-html-the-cthulhu-way/