@cosmicexplorer 2020-06-23
First off, should I have created an issue for this instead of a draft PR first?

Second: Yes! Everything you said is totally correct! One thing that I find is often overlooked is that regular expressions + conditional logic and/or state can create a parser that can parse languages equivalent to CFGs and above. This is how emacs's [`js2-mode`](https://www.emacswiki.org/emacs/Js2Mode) is able to parse javascript to provide semantic-aware highlighting.

The above bit about linguistic power with conditional logic is what wikipedia would call "original research", but I believe it is true. I am actually also working on publishing a novel parsing algorithm which is highly parallelizable and fully general in my copious free time!

The one concern I have for this being feasible long-term is that html parsers like html5lib will handle broken HTML, which would be hard to replicate, *but*:
1. This feature can be kept as an unstable option for a while, with the explicit requirement that all HTML fed to pip must be well-formed (this is easier to ensure if you're resolving against your own repos instead of PyPI).
2. I think this regex parser will also be able to handle broken html (which is not a CFG), because adding conditional logic to regex matching allows us to achieve even greater linguistic power than a CFG parser could.

If this was able to pass the same test suite as the html5lib parser, and there were many more tests added for things like broken HTML, could this feature ever be considered for merging (even under an unstable flag)?