@pfmoore 2020-06-23
One concern I'd have is maintainability. If we were to adopt this code, it would in future be supported, and quite likely modified, by people who are *not* well-versed in the intricacies and risks of correctly parsing (possibly broken) HTML code. So there's a very real chance of bitrot. And describing this algorithm as "original research" doesn't do much to alleviate that concern 🙂 

One of the maybe less obvious advantages of pip relying on external libraries (even if we have to vendor them) is that we delegate the expertise in the respective subject areas onto those downstream library maintainers. How viable would it be to split this code out into a reusable library? Maybe a dedicated PEP 503 parser? Then pip could relatively safely vendor that package *without* taking on the maintenance of that code.

A 12% speed increase is not something we should ignore lightly, but speed is very definitely only one consideration out of many here.