@pradyunsg 2020-06-23
FWIW, there is a dedicated PEP 503 parser -- https://github.com/brettcannon/mousebender. Adopting it is blocked on dropping Python 2 support.