@cosmicexplorer 2020-06-24
@pfmoore thank you for your thoughtful and detailed response! :D

> So there's a very real chance of bitrot. And describing this algorithm as "original research" doesn't do much to alleviate that concern 🙂

Yep! I don't think this idea is really baked enough to work out right now. Thank you so much for talking it through with me!! ^_^

@pradyunsg: that dedicated parser library sounds like it would be a great place to focus my efforts in line with @pfmoore's comment. I do think a dedicated parser library is exactly where I would expect this kind of performance-sensitive code to be.